feat: add deserializer for u128
For some endpoint of the starknet-server, we do want
to accept u128 values for range verification.
However, by default the u128 types only translates
to json number.
This commit introduces a U128HexOrDec deserializer
that allows the retrieval of a u128 value from a string
(hexadecimal or decimal string) without compromising
the u128 range check